+++ routes/satelliteCampus.js
@@ -2,29 +2,27 @@ const express = require("express");
 const router = express.Router();
 const SatelliteCampus = require("../models/satelliteCampus");
 const isNil = require("lodash/isNil");
+1;
 const isEmpty = require("lodash/isEmpty");
 
 // @route   GET api/user
 // @desc    Get All SatelliteCampus
 // @access  Public
 router.get("/", async (req, res) => {
-  const condition = !isNil(req.query.condition) ? JSON.parse(req.query.condition) : {};
+  const condition = !isNil(req.query.condition)
+    ? JSON.parse(req.query.condition)
+    : {};
   if (isNil(condition.deletedAt)) {
-      condition.deletedAt = {
-          $exists: false
-      }
+    condition.deletedAt = {
+      $exists: false,
+    };
   }
   try {
     const getAllSatelliteCampus = await SatelliteCampus.find(condition);
-    res.json({
-      dbRes: getAllSatelliteCampus,
-      isSuccess: true
-    });
-  } catch (error) {
-    res.json({
-      dbRes: error.message,
-      isSuccess: false
-    });
+    res.json(getAllSatelliteCampus);
+  } catch ({ message: errMessage }) {
+    const message = errMessage ? errMessage : UNKNOWN_ERROR_OCCURED;
+    res.status(500).json(message);
   }
 });
 
@@ -36,17 +34,17 @@ router.get("/:id", async (req, res) => {
     const getSatelliteCampus = await SatelliteCampus.findById({
       _id: req.params.id,
       deletedAt: {
-        $exists: false
-      }
+        $exists: false,
+      },
     });
     res.json({
       dbRes: getSatelliteCampus,
-      isSuccess: true
+      isSuccess: true,
     });
   } catch (error) {
     res.json({
       dbRes: error.message,
-      isSuccess: false
+      isSuccess: false,
     });
   }
 });
@@ -64,8 +62,19 @@ router.post("/", async (req, res) => {
   const currencySymbol = req.body.currencySymbol;
   const city = req.body.city;
   const state = req.body.state;
-  const address = req.body.address
-  if (!isNil(campusName) && !isNil(mainCampusId) && !isNil(schoolName) && !isNil(email) && !isNil(mobileNumber) && !isNil(currency) && !isNil(currencySymbol) && !isNil(city) && !isNil(state) && !isNil(address)) {
+  const address = req.body.address;
+  if (
+    !isNil(campusName) &&
+    !isNil(mainCampusId) &&
+    !isNil(schoolName) &&
+    !isNil(email) &&
+    !isNil(mobileNumber) &&
+    !isNil(currency) &&
+    !isNil(currencySymbol) &&
+    !isNil(city) &&
+    !isNil(state) &&
+    !isNil(address)
+  ) {
     const newSatelliteCampus = new SatelliteCampus({
       campusName,
       mainCampusId,
@@ -79,38 +88,27 @@ router.post("/", async (req, res) => {
       address,
       createdAt,
       updatedAt,
-      deletedAt
+      deletedAt,
     });
     try {
       const getSatelliteCampus = await SatelliteCampus.find({
         campusName,
         deletedAt: {
-          $exists: false
-        }
+          $exists: false,
+        },
       });
       if (getSatelliteCampus.length === 0) {
         const createSatelliteCampus = await newSatelliteCampus.save();
-        res.json({
-          dbRes: createSatelliteCampus,
-          isSuccess: true
-        });
+        res.json(createSatelliteCampus);
       } else {
-        res.json({
-          dbRes: "Campus name must be unique",
-          isSuccess: false
-        });
+        res.status(500).json("Campus name must be unique");
       }
-    } catch (error) {
-      res.json({
-        dbRes: error.message,
-        isSuccess: false
-      });
+    } catch ({ message: errMessage }) {
+      const message = errMessage ? errMessage : UNKNOWN_ERROR_OCCURED;
+      res.status(500).json(message);
     }
   } else {
-    res.json({
-      dbRes: "Required values are either invalid or empty",
-      isSuccess: false
-    });
+    res.status(500).json("Required values are either invalid or empty");
   }
 });
 
@@ -118,62 +116,76 @@ router.post("/", async (req, res) => {
 // @desc    Update A SatelliteCampus
 // @access  Private
 router.put("/:id", async (req, res) => {
-    const campusName = req.body.campusName;
-    const mainCampusId = req.body.mainCampusId;
-    const schoolName = req.body.schoolName;
-    const email = req.body.email;
-    const mobileNumber = req.body.mobileNumber;
-    const currency = req.body.currency;
-    const currencySymbol = req.body.currencySymbol;
-    const city = req.body.city;
-    const state = req.body.state;
-    const address = req.body.address
-  if (!isNil(campusName) && !isNil(mainCampusId) && !isNil(schoolName) && !isNil(email) && !isNil(mobileNumber) && !isNil(currency) && !isNil(currencySymbol) && !isNil(city) && !isNil(state) && !isNil(address)) {
+  const campusName = req.body.campusName;
+  const mainCampusId = req.body.mainCampusId;
+  const schoolName = req.body.schoolName;
+  const email = req.body.email;
+  const mobileNumber = req.body.mobileNumber;
+  const currency = req.body.currency;
+  const currencySymbol = req.body.currencySymbol;
+  const city = req.body.city;
+  const state = req.body.state;
+  const address = req.body.address;
+  if (
+    !isNil(campusName) &&
+    !isNil(mainCampusId) &&
+    !isNil(schoolName) &&
+    !isNil(email) &&
+    !isNil(mobileNumber) &&
+    !isNil(currency) &&
+    !isNil(currencySymbol) &&
+    !isNil(city) &&
+    !isNil(state) &&
+    !isNil(address)
+  ) {
     try {
       const getSatelliteCampus = await SatelliteCampus.find({
         campusName,
         deletedAt: {
-          $exists: false
-        }
+          $exists: false,
+        },
       });
       if (getSatelliteCampus.length === 0) {
-        const updateSatelliteCampus = await SatelliteCampus.findByIdAndUpdate(req.params.id, {
-          $set: {
-            campusName,
-            mainCampusId,
-            schoolName,
-            email,
-            mobileNumber,
-            currency,
-            currencySymbol,
-            city,
-            state,
-            address,
-            createdAt,
-            updatedAt: Date.now,
-            deletedAt
-          },
-        });
+        const updateSatelliteCampus = await SatelliteCampus.findByIdAndUpdate(
+          req.params.id,
+          {
+            $set: {
+              campusName,
+              mainCampusId,
+              schoolName,
+              email,
+              mobileNumber,
+              currency,
+              currencySymbol,
+              city,
+              state,
+              address,
+              createdAt,
+              updatedAt: Date.now,
+              deletedAt,
+            },
+          }
+        );
         res.json({
           dbRes: updateSatelliteCampus,
-          isSuccess: true
+          isSuccess: true,
         });
       } else {
         res.json({
           dbRes: "Campus name must be unique",
-          isSuccess: false
+          isSuccess: false,
         });
       }
     } catch (error) {
       res.json({
         dbRes: error.message,
-        isSuccess: false
+        isSuccess: false,
       });
     }
   } else {
     res.json({
       dbRes: "Required values are either invalid or empty",
-      isSuccess: false
+      isSuccess: false,
     });
   }
 });
@@ -185,25 +197,20 @@ router.patch("/:id", async (req, res) => {
   const condition = req.body;
   if (!isEmpty(condition)) {
     try {
-        const updateSatelliteCampus = await SatelliteCampus.findByIdAndUpdate(req.params.id, {
+      const updateSatelliteCampus = await SatelliteCampus.findByIdAndUpdate(
+        req.params.id,
+        {
           $set: condition,
           updatedAt: Date.now(),
-        });
-        res.json({
-          dbRes: updateSatelliteCampus,
-          isSuccess: true
-        });
-    } catch (error) {
-      res.json({
-        dbRes: error.message,
-        isSuccess: false
-      });
+        }
+      );
+      res.json(updateSatelliteCampus);
+    } catch ({ message: errMessage }) {
+      const message = errMessage ? errMessage : UNKNOWN_ERROR_OCCURED;
+      res.status(500).json(message);
     }
   } else {
-    res.json({
-      dbRes: "Course Cannot be found",
-      isSuccess: false
-    });
+    res.status(500).json("Course Cannot be found");
   }
 });
 
@@ -215,30 +222,25 @@ router.delete("/:id", async (req, res) => {
     const getSatelliteCampus = await SatelliteCampus.find({
       _id: req.params.id,
       deletedAt: {
-        $exists: false
-      }
+        $exists: false,
+      },
     });
     if (getSatelliteCampus.length > 0) {
-      const deleteSatelliteCampus = await SatelliteCampus.findByIdAndUpdate(req.params.id, {
-        $set: {
-          deletedAt: Date.now(),
-        },
-      });
-      res.json({
-        dbRes: deleteSatelliteCampus,
-        isSuccess: true
-      });
+      const deleteSatelliteCampus = await SatelliteCampus.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            deletedAt: Date.now(),
+          },
+        }
+      );
+      res.json(deleteSatelliteCampus);
     } else {
-      res.json({
-        dbRes: "Campus is already deleted",
-        isSuccess: false
-      });
+      res.status(500).json("Campus is already deleted");
     }
-  } catch (error) {
-    res.json({
-      dbRes: error.message,
-      isSuccess: false
-    });
+  } catch ({ message: errMessage }) {
+    const message = errMessage ? errMessage : UNKNOWN_ERROR_OCCURED;
+    res.status(500).json(message);
   }
 });